# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NexusPay API
  version: v1
  description: >-
    Curated OpenAPI for the merchant integration surface (payments, refunds,
    payment sessions, webhook endpoints/deliveries, API keys).


    This document is hand-maintained for readability and TypeScript codegen. The
    live `GET /v1/api-docs` (springdoc) is the complete authoritative spec —
    snapshot it with `curl localhost:8090/v1/api-docs -o openapi.json` and
    generate types with `openapi-typescript`.


    Conventions:
      - All monetary amounts are **integers in minor units** (e.g. cents: 5000 = $50.00).
      - Timestamps are ISO-8601 (`Instant`) EXCEPT the webhook envelope `created`
        field, which is **epoch seconds** (see `WebhookEvent`).
      - Response wire keys are **snake_case** for every DTO EXCEPT the SDK
        `ConfirmResponse` (camelCase, pinned `@JsonProperty`), which is the
        `/v1/checkout/confirm` body consumed by `@nexus-pay/js` and is not part of
        this merchant-server surface.
      - `mode` / `livemode` are SERVER-DERIVED from the authenticated key's
        `is_live` flag (test for `sk_test_`), never read from the request body.
servers:
  - url: http://localhost:8090
    description: Local dev (lite stack)
...
paths:
...
  /v1/test/clock:
    put:
      tags: [Test Helpers]
      summary: Freeze the test clock (TEST MODE ONLY)
      description: >-
        TEST-MODE ONLY. Reachable ONLY with an `sk_test_` key — a LIVE key gets
        404 (no oracle). Scope: `test:write`. Freezes the caller tenant's TEST
        clock at `now` (an ISO-8601 instant), so the `created_at` stamped on
        subsequently TEST-created payment/refund artifacts is DETERMINISTIC.
        That frozen `created_at` is CONSISTENT across every read of the same
        artifact: the create response, the `GET /v1/payments` | `/v1/refunds`
        LIST (and its ordering), AND the `GET /v1/payments/{id}` | `/v1/refunds/{id}`
        single-retrieve — there is no real-time split between them.
        HONEST SCOPE: this controls ONLY that `created_at` (and, because the
        read-model inherits it, the `GET /v1/payments` | `/v1/refunds` LIST
        ORDERING). It does NOT control mandate expiry, idempotency-key TTL,
        webhook-delivery retry/backoff, api-key expiry, the read-model
        `updated_at`, the synthesized webhook envelope timestamp, or ANY
        live-rail behaviour — a LIVE charge's timestamp is unaffected. An
        unparseable/missing `now` is rejected `400`; a far-future instant is
        allowed (test freedom). Per-tenant: a tenant can set/read/clear ONLY its
        own clock.
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestClockRequest'
      responses:
        '200':
          description: The clock was frozen for the caller tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestClockResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Endpoint not reachable with this key (live key in test-only surface)
    get:
      tags: [Test Helpers]
      summary: Read the test-clock state (TEST MODE ONLY)
      description: >-
        TEST-MODE ONLY (`sk_test_`; LIVE -> 404). Scope: `test:write`. Returns
        the caller tenant's clock state: `now` is the frozen instant if set, else
        real time; `frozen` is true when a clock is set. Same HONEST SCOPE caveat
        as `PUT` — controls ONLY artifact `created_at`, not expiry/TTL/retry/live.
      security:
        - apiKey: []
      responses:
        '200':
          description: The caller tenant's clock state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestClockResponse'
        '404':
          description: Endpoint not reachable with this key (live key in test-only surface)
    delete:
      tags: [Test Helpers]
      summary: Clear the test clock -> revert to real time (TEST MODE ONLY)
      description: >-
        TEST-MODE ONLY (`sk_test_`; LIVE -> 404). Scope: `test:write`. Clears the
        caller tenant's clock so subsequently TEST-created artifacts are stamped
        with real time again. Returns `{ now: <real time>, frozen: false }`.
      security:
        - apiKey: []
      responses:
        '200':
          description: The clock was cleared; reverted to real time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestClockResponse'
        '404':
          description: Endpoint not reachable with this key (live key in test-only surface)
...
components:
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: >-
        Merchant API key. Send as `Authorization: Bearer sk_test_…` (or
        `sk_live_…` in production). Validated by ApiKeyAuthenticationFilter.
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Short-lived checkout session token (JWT) for `/v1/checkout/**` endpoints
        (browser SDK). Send as `Authorization: Bearer <session-jwt>`.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Keycloak OIDC access token (JWT) for dashboard/back-office endpoints.
        Send as `Authorization: Bearer <oidc-jwt>`.
  schemas:
...
    ApiErrorResponse:
      type: object
      description: 'INT-2 error envelope: { error: { type, code, message, request_id } }.'
      required: [error]
      properties:
        error:
          type: object
          required: [type, code, message, request_id]
          properties:
            type:
              type: string
              enum:
                - validation_error
                - not_found
                - unauthorized
                - forbidden
                - conflict
                - payment_error
                - rate_limit_error
                - internal_error
                - session_error
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Safe, human-readable message (never leaks internals).
            request_id:
              type: string
              description: Correlation id (echoes X-Request-Id); always present.
...
    TestClockRequest:
      type: object
      required: [now]
      description: Body for freezing the test clock (`PUT /v1/test/clock`).
      properties:
        now:
          type: string
          format: date-time
          description: >-
            The ISO-8601 instant to freeze the clock at. Stamped as the
            `created_at` on subsequently TEST-created payment/refund artifacts. An
            unparseable/missing value is rejected `400`; a far-future instant is
            allowed.
          example: '2026-01-01T00:00:00Z'
    TestClockResponse:
      type: object
      description: >-
        The caller tenant's test-clock state. A frozen clock controls ONLY the
        `created_at` on TEST-created payment/refund artifacts (and their list
        ordering) — NOT expiry, idempotency TTL, webhook retry/backoff,
        `updated_at`, the webhook envelope timestamp, or any live-rail behaviour.
      required: [now, frozen]
      properties:
        now:
          type: string
          format: date-time
          description: >-
            The instant the clock resolves to now — the frozen instant if frozen,
            else real time.
          example: '2026-01-01T00:00:00Z'
        frozen:
          type: boolean
          description: True when a frozen clock is set for the tenant; false = real time.
          example: true
...